requirements.txt
streamlit
google-generativeai
dotenv-python
langchain
PyPDF2
chromadb
faiss-cpu
pdf2image
sqlite3